initial commit
The intent of this repo is to mimic the steps one might take when solving an optimization problem.

The very first step is to have some idea of what you are even optimizing.
Thus, we need some model of what a solution looks like, and some way to score that solution.